backend/pyproject.toml
[tool.poetry]
name = "langgraph_agents"
version = "0.1.0"
description = ""
authors = ["Rahul Bhatia <rahul.bhatia023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
cohere = "5.9.2"
langgraph = "^0.2.21"
langchain-openai = "^0.2.0"
langchain = "^0.3.0"
black = "^24.8.0"
langchain-community = "^0.3.0"
praw = "^7.7.1"
semantic-router = "^0.0.61"
langchain-ollama = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
